dev-requirements.txt
aiohttp>=1.0
namedlist>=1.6
python-dateutil>=2.5
simplejson
pytest
pytest-mock
pytest-asyncio
